+++ HRealEngine/src/HRealEngine/Renderer/Renderer2D.cpp
@@ -589,7 +589,7 @@ namespace HRealEngine
         double fsScale = 1.0 / (metrics.ascenderY - metrics.descenderY);
         double y = 0.0;
 
-        const float spaceGlyphAdvance = fontGeometry.getGlyph(' ')->getAdvance();
+        const float spaceGlyphAdvance = static_cast<float>(fontGeometry.getGlyph(' ')->getAdvance());
 
         for (size_t i = 0; i < string.size(); i++)
         {
